gpt-line-bot/python/requirements.txt
functions-framework==3.*
line-bot-sdk==3.7.0
openai==1.11.0
google-api-python-client==2.116.0
google-cloud-logging==3.9.0
google-cloud-firestore==2.14.0
arxiv==2.1.0
xmltodict==0.13.0
langchain==0.1.5
langchainhub==0.1.14
langchain-openai==0.0.5
numexpr==2.9.0
